Add loader + ConfirmDialog

+++ pages/admin/components/teams/TeamCard.tsx
@@ -1,8 +1,16 @@
-import { Box, VStack, Image, IconButton, useDisclosure } from '@chakra-ui/react'
+import {
+  Box,
+  VStack,
+  Image,
+  IconButton,
+  useDisclosure,
+  Grid
+} from '@chakra-ui/react'
 import { Types } from 'mongoose'
 import { FiXCircle, FiEdit2 } from 'react-icons/fi'
 import { ITeam } from '../../../../server/db/models/teamModel'
 import UpdateTeamModal from './modal/UpdateTeamModal'
+import ConfirmAlertDialog from '../../../../common/components/ConfirmAlertDialog'
 
 type TeamCardProps = {
   team: ITeam
@@ -11,7 +19,16 @@ type TeamCardProps = {
 }
 
 function TeamCard({ team, deleteTeam, updateTeam }: TeamCardProps) {
-  const { isOpen, onOpen, onClose } = useDisclosure()
+  const {
+    isOpen: isModalOpen,
+    onOpen: onModalOpen,
+    onClose: onCloseModal
+  } = useDisclosure()
+  const {
+    isOpen: isDialogOpen,
+    onOpen: onOpenDialog,
+    onClose: onCloseDialog
+  } = useDisclosure()
 
   return (
     <Box
@@ -22,29 +39,38 @@ function TeamCard({ team, deleteTeam, updateTeam }: TeamCardProps) {
       width="100%"
       boxShadow="lg"
     >
-      <Box>
-        <IconButton
-          fontSize={30}
-          icon={<FiEdit2 />}
-          aria-label="edit team"
-          bg="white"
-          onClick={onOpen}
-        />
-        <UpdateTeamModal
-          onSave={updateTeam}
-          onClose={onClose}
-          isOpen={isOpen}
-          team={team}
-        />
-        <IconButton
-          fontSize={30}
-          icon={<FiXCircle />}
-          aria-label="delete team"
-          bg="white"
-          color="red.500"
-          onClick={() => deleteTeam(team._id)}
-        />
-      </Box>
+      <Grid templateColumns="1fr auto" gap={4} alignItems="center">
+        <Box>
+          <IconButton
+            fontSize={30}
+            icon={<FiEdit2 />}
+            aria-label="edit team"
+            bg="white"
+            onClick={onModalOpen}
+          />
+          <UpdateTeamModal
+            onSave={updateTeam}
+            onClose={onCloseModal}
+            isOpen={isModalOpen}
+            team={team}
+          />
+        </Box>
+        <Box justifySelf="flex-end">
+          <IconButton
+            fontSize={30}
+            icon={<FiXCircle />}
+            aria-label="delete team"
+            bg="white"
+            color="red.500"
+            onClick={onOpenDialog}
+          />
+          <ConfirmAlertDialog
+            isOpen={isDialogOpen}
+            onClose={onCloseDialog}
+            confirmAction={() => deleteTeam(team._id)}
+          />
+        </Box>
+      </Grid>
       <VStack spacing={4}>
         <Image
           src={`/assets/${team.logoURI}`}
